+++ include/datamanagement/datamanagement.hpp
@@ -1,8 +1,25 @@
-#ifndef DATAMANAGEMENT_HPP_
-#define DATAMANAGEMENT_HPP_
+////////////////////////////////////////////////////////////////////////////////
+// File: datamanagement.hpp                                                   //
+// Project: datamanagement                                                    //
+// Created Date: Th Feb 2025                                                  //
+// Author: Matthew Carroll                                                    //
+// -----                                                                      //
+// Last Modified: Thu Feb 20 2025                                             //
+// Modified By: Matthew Carroll                                               //
+// -----                                                                      //
+// Copyright (c) 2025 Syndemics Lab at Boston Medical Center                  //
+// -----                                                                      //
+// HISTORY:                                                                   //
+// Date      	By	Comments                                                  //
+// ----------	---	--------------------------------------------------------- //
+////////////////////////////////////////////////////////////////////////////////
 
-#include <datamanagement/modeldata/ModelData.hpp>
-#include <datamanagement/source/Config.hpp>
-#include <datamanagement/source/TableSource.hpp>
+#ifndef DATAMANAGEMENT_DATAMANAGEMENT_HPP_
+#define DATAMANAGEMENT_DATAMANAGEMENT_HPP_
 
-#endif
+#include <datamanagement/modeldata/model_data.hpp>
+#include <datamanagement/source/config.hpp>
+#include <datamanagement/source/csv_source.hpp>
+#include <datamanagement/source/db_source.hpp>
+
+#endif // DATAMANAGEMENT_DATAMANAGEMENT_HPP_
